Added all logic to deal with adding an item to cart in local storage to useSaveCartOnLocalStorage
