+++ src/ProductDetails.tsx
@@ -3,32 +3,33 @@ import { useTranslation } from "react-i18next";
 import ITEM_LIST from "./lists/ITEM_LIST.ts";
 import { ItemDetailsModel } from "./models";
 import { Link, useSearchParams } from "react-router-dom";
-import { JSONCartModel } from "./models/cartModel.ts";
+// import { JSONCartModel } from "./models/cartModel.ts";
 import { useSaveCartOnLocalStorage } from "./utils/useSaveCartOnLocalStorage.tsx";
-import { getCartFromLocalStorage } from "./utils/getCartFromLocalStorage.tsx";
+// import { getCartFromLocalStorage } from "./utils/getCartFromLocalStorage.tsx";
 
 const Product = () => {
   const { t } = useTranslation("productDetails");
   const [param] = useSearchParams();
   const productIdParam = Number(param.get("ref"));
+  const { saveItemToCart } = useSaveCartOnLocalStorage();
 
   const selectedItem = ITEM_LIST.find((item: ItemDetailsModel) => {
     return item.key === productIdParam;
   });
 
   const addItemToCart = (key: number) => {
-    // Adds a default quantity of one as quantities will be dealt with in the cart site
-    const JSONCart: JSONCartModel = {
-      itemId: key,
-      quantity: 1,
-    };
-
-    // Gets previous cart, pushes new element to cart (JSONCart), and saves new cart on local storage
-    const cart: JSONCartModel[] = getCartFromLocalStorage();
-
-    cart.push(JSONCart);
-
-    useSaveCartOnLocalStorage(cart);
+    // // Adds a default quantity of one as quantities will be dealt with in the cart site
+    // const JSONCart: JSONCartModel = {
+    //   itemId: key,
+    //   quantity: 1,
+    // };
+    //
+    // // Gets previous cart, pushes new element to cart (JSONCart), and saves new cart on local storage
+    // const cart: JSONCartModel[] = getCartFromLocalStorage();
+    //
+    // cart.push(JSONCart);
+    //
+    saveItemToCart(key);
   };
 
   return (
